move: 각 유스케이스들 하나의 파일로 이동
