--- app/src/main/java/com/egobook/app/domain/usecase/authusecase/GoogleAutoLogin.kt
@@ -1,13 +1,0 @@
-package com.egobook.app.domain.usecase.authusecase
-
-import com.egobook.app.domain.repository.auth.AuthRepository
-import javax.inject.Inject
-
-class GoogleAutoLogin @Inject constructor(
-    private val repository: AuthRepository
-) {
-    suspend operator fun invoke(): Result<Unit> {
-        return repository.refreshAccessToken() //액세스토큰 갱신
-    }
-
-}
