+++ app/src/main/java/com/egobook/app/domain/usecase/diaryusecase/DiaryUseCases.kt
@@ -1,8 +1,14 @@
 package com.egobook.app.domain.usecase.diaryusecase
 
+import com.egobook.app.domain.model.Diary
+import com.egobook.app.domain.model.DiaryType
+import com.egobook.app.domain.repository.DiaryRepository
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.map
+import java.time.LocalDateTime
 import javax.inject.Inject
 
-// 의존성 주입을 쉽게 하기 위한 래퍼 클래스. DiaryFragment DiaryWriteFragment, DiaryCheckFragment에서의 유스케이스 모음
+// 의존성 주입을 쉽게 하기 위한 래퍼 클래스.
 data class DiaryUseCases @Inject constructor (
     val getDiaries: GetDiaries,
     val getDiary: GetDiary,
@@ -10,3 +16,84 @@ data class DiaryUseCases @Inject constructor (
     val updateDiary: UpdateDiary,
     val deleteDiary: DeleteDiary
 )
+
+//각 유스케이스들 정의
+
+class GetDiaries @Inject constructor(
+    private val repository: DiaryRepository
+) {
+    operator fun invoke(
+        selectedDate: LocalDateTime = LocalDateTime.now(),
+        types: Set<DiaryType>? = null // null = 전체 탭
+    ): Flow<List<Diary>> {
+        return repository.getDiaries()
+            .map { diaries ->
+                diaries
+                    .filter { diary ->
+                        val isSameDate = diary.createdAt.year == selectedDate.year &&
+                                diary.createdAt.month == selectedDate.month &&
+                                diary.createdAt.dayOfMonth == selectedDate.dayOfMonth
+                        val isCorrectType = types == null || diary.types.any { it in types }
+
+                        isSameDate && isCorrectType
+                    }
+                    .sortedByDescending { it.writtenAt }
+            }
+    }
+}
+
+class GetDiary @Inject constructor(
+    private val repository: DiaryRepository
+) {
+    suspend operator fun invoke(id: Long): Result<Diary?> {
+        return repository.getDiaryById(id)
+    }
+}
+
+class AddDiary @Inject constructor(
+    private val repository: DiaryRepository
+) {
+    suspend operator fun invoke(
+        content: String,
+        types: Set<DiaryType>,
+        emotionLevel: Int?, // 1~5 사이의 감정 레벨
+        createdAt: LocalDateTime // 일기가 귀속될 날짜
+    ): Result<Diary> {
+        return repository.addDiary(
+            content = content,
+            types = types,
+            emotionLevel = emotionLevel,
+            createdAt = createdAt
+        )
+    }
+}
+
+class UpdateDiary @Inject constructor(
+    private val repository: DiaryRepository
+) {
+    suspend operator fun invoke(
+        id: Long,
+        content: String,
+        types: Set<DiaryType>,
+        emotionLevel: Int?
+    ): Result<Diary> {
+        return repository.updateDiary(
+            id = id,
+            content = content,
+            types = types,
+            emotionLevel = emotionLevel
+        )
+    }
+}
+
+class DeleteDiary @Inject constructor(
+    private val repository: DiaryRepository
+) {
+    suspend operator fun invoke(id: Long): Result<Unit> {
+        return repository.deleteDiaryById(id)
+    }
+}
+
+
+
+
