+++ app/src/main/res/layout/activity_main.xml
@@ -59,6 +59,17 @@
 
         </eightbitlab.com.blurview.BlurView>
 
+        <ProgressBar
+            android:id="@+id/loading_progress"
+            style="@style/LoadingProgressBarStyle"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:indeterminate="true"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent" />
+
     </androidx.constraintlayout.widget.ConstraintLayout>
 
     <androidx.fragment.app.FragmentContainerView
