+++ app/src/main/res/layout/fragment_home.xml
@@ -171,4 +171,5 @@
         android:orientation="vertical"
         app:layout_constraintGuide_begin="16dp"/>
 
+
 </androidx.constraintlayout.widget.ConstraintLayout>
