+++ app/src/main/res/values/styles.xml
@@ -121,7 +121,6 @@
         <item name="android:textSize">12sp</item>
         <item name="android:fontFamily">@font/arita_semibold</item>
         <item name="android:gravity">center</item>
-        <item name="android:gravity">center</item>
         <item name="android:letterSpacing">-0.03</item>
         <item name="android:textAllCaps">false</item>
 
@@ -183,6 +182,14 @@
         <item name="android:layout_height">wrap_content</item>
         <item name="android:fontFamily">@font/arita_semibold</item>
         <item name="android:textColor">@color/cos_black</item>
+    </style>
+
+    <!-- 로딩 ProgressBar 스타일 -->
+    <style name="LoadingProgressBarStyle" parent="Widget.AppCompat.ProgressBar">
+        <item name="android:indeterminateTint">@color/emotion_very_happy</item>
+        <item name="android:indeterminateTintMode">src_atop</item>
+        <item name="android:minWidth">48dp</item>
+        <item name="android:minHeight">48dp</item>
     </style>
